SearchSpace commit 1

+++ Client/ImageLoader.cpp
@@ -83,7 +83,7 @@ Image ImageLoader::loadImage(std::string filePath) {
     int row = 0;
     int col = 0;
 
-    for (int i = 0; i < output.size(); i = i + 4) {
+    for (size_t i = 0; i < output.size(); i = i + 4) {
         int r = output[i];
         int g = output[i+1];
         int b = output[i+2];
